feat: removed retry
